requirements.txt
adafruit-blinka
adafruit-circuitpython-ads1x15
gas-detection
python-dotenv
requests
